@@ -10,8 +10,8 @@
   "dependencies": {
     "babel-plugin-react-native-web": "0.18.12",
     "next": "^12.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-native-web": "0.18.12"
   },
   "devDependencies": {
