@@ -31,8 +31,8 @@
     "styleq": "^0.1.2"
   },
   "peerDependencies": {
-    "react": "^17.0.2 || ^18.0.0",
-    "react-dom": "^17.0.2 || ^18.0.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "author": "Nicolas Gallagher",
   "license": "MIT",
