@@ -7,9 +7,8 @@
 
 import ActivityIndicator from '..';
 import React from 'react';
-import { act } from 'react-dom/test-utils';
 import { createEventTarget } from 'dom-event-testing-library';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 
 describe('components/ActivityIndicator', () => {
   describe('prop "accessibilityLabel"', () => {
