@@ -1,8 +1,7 @@
 import Button from '..';
 import React from 'react';
-import { act } from 'react-dom/test-utils';
 import { createEventTarget } from 'dom-event-testing-library';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 
 describe('components/Button', () => {
   test('prop "accessibilityLabel"', () => {
