@@ -7,9 +7,8 @@
 
 import CheckBox from '../';
 import React from 'react';
-import { act } from 'react-dom/test-utils';
 import { createEventTarget } from 'dom-event-testing-library';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 
 function findCheckbox(container) {
   return container.firstChild.querySelector('input');
