@@ -7,9 +7,8 @@
 
 import React from 'react';
 import TextInput from '..';
-import { act } from 'react-dom/test-utils';
 import { createEventTarget } from 'dom-event-testing-library';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 
 function findInput(container) {
   return container.querySelector('input');
