@@ -8,9 +8,8 @@
 import React from 'react';
 import View from '../';
 import StyleSheet from '../../StyleSheet';
-import { act } from 'react-dom/test-utils';
 import { createEventTarget } from 'dom-event-testing-library';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 
 describe('components/View', () => {
   test('default', () => {
