@@ -8,53 +8,11 @@
  */
 
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import * as ReactDOMServer from 'react-dom/server';
-import { act } from 'react-dom/test-utils';
+import { act, render } from '@testing-library/react';
 import { createEventTarget } from 'dom-event-testing-library';
 import createEventHandle from '..';
 
-function createRoot(rootNode) {
-  return {
-    render(element) {
-      ReactDOM.render(element, rootNode);
-    }
-  };
-}
-
 describe('create-event-handle', () => {
-  let root;
-  let rootNode;
-
-  beforeEach(() => {
-    rootNode = document.createElement('div');
-    document.body.appendChild(rootNode);
-    root = createRoot(rootNode);
-  });
-
-  afterEach(() => {
-    root.render(null);
-    document.body.removeChild(rootNode);
-    rootNode = null;
-    root = null;
-  });
-
-  test('can render correctly using ReactDOMServer', () => {
-    const listener = jest.fn();
-    const targetRef = React.createRef();
-    const addClickListener = createEventHandle('click');
-
-    function Component() {
-      React.useEffect(() => {
-        return addClickListener(targetRef.current, listener);
-      });
-      return <div ref={targetRef} />;
-    }
-
-    const output = ReactDOMServer.renderToString(<Component />);
-    expect(output).toBe('<div data-reactroot=""></div>');
-  });
-
   describe('createEventTarget()', () => {
     test('event dispatched on target', () => {
       const listener = jest.fn();
@@ -68,9 +26,7 @@ describe('create-event-handle', () => {
         return <div ref={targetRef} />;
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const target = createEventTarget(targetRef.current);
 
@@ -113,9 +69,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const parent = createEventTarget(parentRef.current);
 
@@ -164,9 +118,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const child = createEventTarget(childRef.current);
 
@@ -196,9 +148,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const text = createEventTarget(childRef.current.firstChild);
 
@@ -221,9 +171,8 @@ describe('create-event-handle', () => {
         return <div ref={targetRef} />;
       }
 
-      act(() => {
-        root.render(<Component target={document} />);
-      });
+      render(<Component target={document} />);
+
       const target = createEventTarget(targetRef.current);
       act(() => {
         target.click();
@@ -244,9 +193,8 @@ describe('create-event-handle', () => {
         return <div ref={targetRef} />;
       }
 
-      act(() => {
-        root.render(<Component target={window} />);
-      });
+      render(<Component target={window} />);
+
       const target = createEventTarget(targetRef.current);
       act(() => {
         target.click();
@@ -267,9 +215,7 @@ describe('create-event-handle', () => {
         return <div ref={targetRef} />;
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       act(() => {
         const event = new CustomEvent('magic-event', { bubbles: true });
@@ -312,9 +258,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const child = createEventTarget(childRef.current);
 
@@ -371,9 +315,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const child = createEventTarget(childRef.current);
 
@@ -416,9 +358,7 @@ describe('create-event-handle', () => {
         );
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const child = createEventTarget(childRef.current);
 
@@ -447,9 +387,7 @@ describe('create-event-handle', () => {
         return <div ref={targetRef} />;
       }
 
-      act(() => {
-        root.render(<Component />);
-      });
+      render(<Component />);
 
       const target = createEventTarget(targetRef.current);
 
