@@ -5,9 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { act } from 'react-dom/test-utils';
+import { act, render } from '@testing-library/react';
 import React, { createRef } from 'react';
-import ReactDOM from 'react-dom';
 import useResponderEvents from '..';
 import { getResponderNode, terminateResponder } from '../ResponderSystem';
 import {
@@ -19,21 +18,7 @@ import {
 } from 'dom-event-testing-library';
 
 describe('useResponderEvents', () => {
-  let container;
-
-  function render(element) {
-    ReactDOM.render(element, container);
-  }
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
   afterEach(() => {
-    render(null);
-    document.body.removeChild(container);
-    container = null;
     // make sure all tests end with the current responder being reset
     terminateResponder();
     // make sure all tests reset state machine tracking pointers on the mock surface
@@ -52,9 +37,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture
       act(() => {
@@ -76,9 +59,7 @@ describe('useResponderEvents', () => {
     };
 
     // render
-    act(() => {
-      render(<Component />);
-    });
+    render(<Component />);
     const target = createEventTarget(targetRef.current);
     const buttons = [1, 2, 3, 4];
     // gesture
@@ -105,9 +86,7 @@ describe('useResponderEvents', () => {
     };
 
     // render
-    act(() => {
-      render(<Component />);
-    });
+    render(<Component />);
     const target = createEventTarget(targetRef.current);
     const acceptedModifierKeys = ['metaKey', 'shiftKey'];
     const ignoredModifierKeys = ['altKey', 'ctrlKey'];
@@ -139,9 +118,7 @@ describe('useResponderEvents', () => {
     };
 
     // render
-    act(() => {
-      render(<Component />);
-    });
+    render(<Component />);
     const target = createEventTarget(targetRef.current);
     // touch gesture
     act(() => {
@@ -186,9 +163,7 @@ describe('useResponderEvents', () => {
     };
 
     // render
-    act(() => {
-      render(<Component />);
-    });
+    render(<Component />);
     const target = createEventTarget(targetRef.current);
     // gesture
     act(() => {
@@ -245,9 +220,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -302,9 +275,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start
       act(() => {
@@ -355,9 +326,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start
       act(() => {
@@ -426,9 +395,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start
       act(() => {
@@ -475,9 +442,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start
       act(() => {
@@ -526,9 +491,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -596,9 +559,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start & move
         act(() => {
@@ -653,9 +614,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start & move
       act(() => {
@@ -707,9 +666,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start & move
       act(() => {
@@ -779,9 +736,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start & move
       act(() => {
@@ -830,9 +785,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start & move
       act(() => {
@@ -882,9 +835,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start & move
         act(() => {
@@ -943,9 +894,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -987,9 +936,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -1038,9 +985,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start
       act(() => {
@@ -1067,9 +1012,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -1134,9 +1077,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start
         act(() => {
@@ -1181,9 +1122,7 @@ describe('useResponderEvents', () => {
             };
 
             // render
-            act(() => {
-              render(<Component />);
-            });
+            render(<Component />);
             const target = createEventTarget(targetRef.current);
             // gesture start & move
             act(() => {
@@ -1228,9 +1167,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture start & end
         act(() => {
@@ -1273,9 +1210,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         // gesture
         act(() => {
@@ -1320,9 +1255,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // gesture start & cancel
       act(() => {
@@ -1359,9 +1292,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const input = createEventTarget(inputRef.current);
       // getSelection is not supported in jest
@@ -1389,9 +1320,7 @@ describe('useResponderEvents', () => {
         };
 
         // render
-        act(() => {
-          render(<Component />);
-        });
+        render(<Component />);
         const target = createEventTarget(targetRef.current);
         const doc = createEventTarget(document);
         // getSelection is not supported in jest
@@ -1437,9 +1366,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const parent = createEventTarget(parentRef.current);
       // gesture start & scroll
@@ -1473,9 +1400,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const doc = createEventTarget(document);
       // gesture start & scroll
@@ -1510,9 +1435,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const sibling = createEventTarget(siblingRef.current);
       // gesture start & scroll
@@ -1541,9 +1464,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const doc = createEventTarget(document);
       // gesture start & blur
@@ -1571,9 +1492,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const win = createEventTarget(window);
       // gesture start & blur
@@ -1607,9 +1526,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const sibling = createEventTarget(siblingRef.current);
       // gesture start & blur
@@ -1640,9 +1557,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // contextmenu sequence includes pointerdown "start"
       act(() => {
@@ -1675,9 +1590,7 @@ describe('useResponderEvents', () => {
       }
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const doc = createEventTarget(document);
       // contextmenu
@@ -1860,9 +1773,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
 
       // gesture start
@@ -2034,9 +1945,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
 
       // gesture start
@@ -2199,9 +2108,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const parent = createEventTarget(parentRef.current);
       const target = createEventTarget(targetRef.current);
 
@@ -2387,9 +2294,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const sibling = createEventTarget(siblingRef.current);
       // gesture start on target
@@ -2530,9 +2435,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const sibling = createEventTarget(siblingRef.current);
       // gesture start and move on target
@@ -2636,9 +2539,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       // first touch
       act(() => {
@@ -2773,9 +2674,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       const parent = createEventTarget(parentRef.current);
       // first touch
@@ -2876,9 +2775,7 @@ describe('useResponderEvents', () => {
       };
 
       // render
-      act(() => {
-        render(<Component />);
-      });
+      render(<Component />);
       const target = createEventTarget(targetRef.current);
       act(() => {
         target.pointerdown({ pointerType });
